add : graph 파일 구조 정리 및 연결
