+++ backend/app/engine/nodes/tools.py
@@ -1,0 +1,4 @@
+from langgraph.prebuilt import ToolNode
+
+supervisor_tools_node = ToolNode(tools=[])
+quiz_tools_node = ToolNode(tools=[])
